requirements.txt
aiogram==3.10.0
google-generativeai
flask